+++ scripts/mcp-tool-pack-baseline.json
@@ -25,7 +25,6 @@
   "dpf_test_kernel_refuse_probe",
   "drive_browser_task",
   "edit_sandbox_file",
-  "evaluate_tool",
   "execute_promotion",
   "generate_code",
   "generate_custom_archetype",
@@ -81,9 +80,7 @@
   "save_phase_handoff",
   "schedule_promotion",
   "schedule_release_bundle",
-  "search_portfolio_context",
   "search_sandbox",
-  "search_tool_marketplace",
   "set_change_disposition",
   "setup_email",
   "size_backlog_item",
